// builds/builds/pZme2DcC/0/gamisodes/gamisodesnft/dapper-wallet-storefront/generated/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  DateTime: any;
  EmailAddress: any;
  JSON: any;
  JSONObject: any;
  PositiveFloat: any;
  PositiveInt: any;
  URL: any;
  UnsignedFloat: any;
  UnsignedInt: any;
  Upload: any;
};
// ...
export type File = {
  /** The MIME content type for this file. */
  contentType?: Maybe<Scalars['String']>;
  /** A unique identifier for this file in the Niftory API. */
  id: Scalars['ID'];
  /** The MD5 hash of this file. */
  md5?: Maybe<Scalars['String']>;
  /** A friendly name for the file. */
  name: Scalars['String'];
  /** The upload state of the file. */
  state: FileState;
  /** The cloud storage URL for this file. If state is GENERATED_UPLOAD_URL, then this url is the presigned URL to upload to. */
  url: Scalars['URL'];
};

/** The upload state of a File. */
export enum FileState {
  /** The file failed to ready. */
  Error = 'ERROR',
  /** Niftory has created a pre-signed URL where the file can be uploaded. */
  GeneratedUploadUrl = 'GENERATED_UPLOAD_URL',
  /** Niftory has created a file entry in the database table. */
  Pending = 'PENDING',
  /** The file is ready for use. */
  Ready = 'READY',
  /** The file has been uploaded to a cloud storage for fast retrieval. */
  UploadedToCloudStorage = 'UPLOADED_TO_CLOUD_STORAGE',
  /** The file (and potentially its corresponding metadata) have been uploaded to IPFS. */
  UploadedToIpfs = 'UPLOADED_TO_IPFS'
}
// ...
export enum NftBlockchainState {
  /** An error occurred with this item's last blockchain operation. */
  Error = 'ERROR',
  /** The item is minted, and are no pending operations on it. */
  Minted = 'MINTED',
  /** The item is being minted. */
  Minting = 'MINTING',
  /** The item has already been transferred to a user's wallet. */
  Transferred = 'TRANSFERRED',
  /** The item is being transferred. */
  Transferring = 'TRANSFERRING',
  /** The item hasn't been minted yet. */
  Unminted = 'UNMINTED'
}
// ...
export enum NftModelBlockchainState {
  /** All NFTs in this model have been transferred to users. */
  Completed = 'COMPLETED',
  /** An error occurred with this item's last blockchain operation. */
  Error = 'ERROR',
  /** The item is minted, and are no pending operations on it. */
  Minted = 'MINTED',
  /** The item is being minted. */
  Minting = 'MINTING',
  /** The item hasn't been minted yet. */
  Unminted = 'UNMINTED'
}
// ...
export type SimpleFile = File & {
  __typename?: 'SimpleFile';
  /** The MIME content type for this file. */
  contentType?: Maybe<Scalars['String']>;
  /** A unique identifier for this file in the Niftory API. */
  id: Scalars['ID'];
  /** The MD5 hash of this file. */
  md5?: Maybe<Scalars['String']>;
  /** A friendly name for the file. */
  name: Scalars['String'];
  /** The upload state of the file. */
  state: FileState;
  /** The cloud storage URL for this file. If state is GENERATED_UPLOAD_URL, then this url is the presigned URL to upload to. */
  url: Scalars['URL'];
};

/** The default rarity levels in the Niftory API. */
export enum SimpleRarityLevel {
  /** The most common NFTs. */
  Common = 'COMMON',
  /** The rarest of the rare NFTs, for the most dedicated collectors. */
  Legendary = 'LEGENDARY',
  /** These are rarer, harder to get and more expensive. */
  Rare = 'RARE'
}

/** Status of this resource for user workflows. */
export enum Status {
  /** "Here you go!" */
  Done = 'DONE',
  /** "I'm just getting started." */
  Draft = 'DRAFT',
  /** "I'm working on it!" */
  InProgress = 'IN_PROGRESS',
  /** "I'll get to it eventually..." */
  ToDo = 'TO_DO'
}

/** The state of an item being transferred. */
export enum TransferState {
  /** The item has been created, but not transferred. */
  Available = 'AVAILABLE',
  /** The item failed to transfer. */
  Error = 'ERROR',
  /** The item is being transferred. */
  InProgress = 'IN_PROGRESS',
  /** The item is reserved for a future transfer. */
  Reserved = 'RESERVED',
  /** The item has been transferred. */
  Success = 'SUCCESS'
}
// ...
export type NftsByWalletQuery = { __typename?: 'Query', nftsByWallet?: { __typename?: 'NFTList', cursor?: string | null, items?: Array<{ __typename?: 'NFT', id: string, blockchainId?: string | null, serialNumber?: number | null, blockchainState: NftBlockchainState, status?: TransferState | null, model?: { __typename?: 'NFTModel', id: string, title: string, description: string, rarity?: SimpleRarityLevel | null, attributes?: any | null, quantity?: any | null, quantityMinted?: any | null, metadata?: any | null, status?: Status | null, state: NftModelBlockchainState, updatedAt?: any | null, blockchainId?: string | null, createdAt: any, content?: { __typename?: 'NFTContent', id: string, files?: Array<{ __typename?: 'NFTFile', contentType?: string | null, id: string, url: any } | null> | null, poster?: { __typename?: 'SimpleFile', url: any, state: FileState, contentType?: string | null, id: string } | null } | null } | null } | null> | null } | null };